This changes is the completed application, which allows users can ask questions and can answer a question real-time.
The application shows the list of questions to the user after logged
in. The user can add a question by click the “+” button on the right
navigation bar. Selects any question will show all answers of that
question, and the user can answers the question real-time as well as
see other users answer real-time